requirements.txt
AdvancedHTTPServer==2.1.0
ajpy==0.0.4
alembic==1.0.0.dev0
aniso8601==4.1.0
asn1crypto==0.24.0
backdoor-factory==0.0.0
basemap==1.2.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
binwalk==2.1.2
blinker==1.4
boltons==18.0.1
Brlapi==0.6.7
Brotli==1.0.7
capstone==3.0.5
certifi==2018.8.24
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.3.7
crcelk==1.3
Creoleparser==0.7.5
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
debtags==2.1
decorator==4.3.0
defusedxml==0.6.0
distro==1.3.0
distro-info==0.21
dnspython==1.16.0
ecdsa==0.13
email-validator==1.0.3
entrypoints==0.3
Flask==1.1.1
Flask-Admin==1.5.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.16.0
Genshi==0.7.3
geoip2==2.9.0
geojson==2.4.1
graphene==2.1.3
graphene-sqlalchemy==2.1.2
graphql-core==2.1
graphql-relay==0.4.5
h11==0.8.1
h2==3.0.1
hashID==3.1.4
hpack==3.0.0
html5lib==1.0.1
httplib2==0.11.3
hyperframe==5.1.0
icalendar==4.0.3
idna==2.6
IPy==0.83
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
jupyter-core==4.4.0
kaitaistruct==0.8
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
ldap3==2.5.1
louis==3.8.0
lxml==4.3.2
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
maxminddb==1.4.1
mitmproxy==4.0.4
msgpack==0.5.6
mysqlclient==1.3.10
nassl==2.1.2
nbformat==4.4.0
numpy==1.16.2
olefile==0.46
paramiko==2.4.2
passlib==1.7.1
patator==0.7
pefile==2018.8.8
Pillow==5.4.1
plotly==3.6.1
pluginbase==1.0.0
ply==3.11
promise==2.2
psycopg2==2.7.7
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.30.4
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyotp==2.2.7
pyparsing==2.2.0
pyperclip==1.6.4
pyproj==1.9.6
pyqtgraph==0.10.0
pyserial==3.4
pyshp==2.1.0
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
pysmi==0.3.2
pysnmp==4.4.6
PySocks==1.6.8
python-apt==1.8.4
python-dateutil==2.7.3
python-debian==0.1.34
python-debianbts==2.8.2
python-editor==1.0.3
python-openid==2.2.5
python-pam==1.8.4
python3-openid==3.1.0
pytz==2019.1
pyxdg==0.25
PyYAML==3.13
reportbug==7.5.2
requests==2.21.0
requests-file==1.4.3
retrying==1.3.3
ruamel.yaml==0.15.34
rule-engine==1.1.0
Rx==1.6.1
scipy==1.1.0
SecretStorage==2.3.1
six==1.12.0
smoke-zephyr==1.4.1
sortedcontainers==2.0.4
soupsieve==1.8
SQLAlchemy==1.3.1
sslyze==2.0.6
tabulate==0.8.2
termcolor==1.1.0
termineter==1.0.4
tld==0.9.1
tls-parser==1.2.1
tornado==5.1.1
traitlets==4.3.2
tzlocal==1.5.1
unattended-upgrades==0.1
urllib3==1.24.1
urwid==2.0.1
wafw00f==0.9.5
wapiti3==3.0.1
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.15.5
wfuzz==2.3.4
Whoosh==2.7.4
wsproto==0.11.0
WTForms==2.2.1
XlsxWriter==1.1.2
yaswfp==0.9.3
